[JAX] Replace uses of jnp.float_, jnp.int_, and jnp.complex_ with float, int, and complex in .astype, jnp.zeros, and jnp.ones.
Right now, this change does nothing except suppress a warning. For example, `jnp.float_` is an alias for `jnp.float64`, and is treated as such by both jax.numpy and numpy. However, JAX warns if `jnp.float64` is passed as a dtype if x64 mode is disabled and truncates the type to `jnp.float32`. `float` acts exactly the same way, except without the warning: it yields a `float32` if x64 mode is disabled, and a `float64` if x64 mode is enabled, with no warning. This change is in preparation for making the warning into an error.

PiperOrigin-RevId: 914813920

+++ distrax/_src/bijectors/bijector_test.py
@@ -28,11 +28,11 @@ class DummyBijector(bijector.Bijector):
 
   def forward_and_log_det(self, x):
     super()._check_forward_input_shape(x)
-    return x, jnp.zeros(x.shape[:-1], jnp.float_)
+    return x, jnp.zeros(x.shape[:-1], float)
 
   def inverse_and_log_det(self, y):
     super()._check_inverse_input_shape(y)
-    return y, jnp.zeros(y.shape[:-1], jnp.float_)
+    return y, jnp.zeros(y.shape[:-1], float)
 
 
 class BijectorTest(parameterized.TestCase):
